pyproject.toml
[project]
name = "nice-app-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ansys-dyna-core>=0.9.0",
    "nicegui>=3.4.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pyvista>=0.46.4",
    "scipy>=1.16.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
